Implemented security, user, rubric, and section

+++ backend/src/main/java/team/projectpulse/rubric/dto/CriterionDto.java
@@ -1,4 +1,5 @@
 package team.projectpulse.rubric.dto;
 
-// Owner: Josh (Person 1)
-public record CriterionDto(Integer id, String name, String description, Double maxScore, Integer sortOrder) {}
+import java.math.BigDecimal;
+
+public record CriterionDto(Integer id, String name, String description, BigDecimal maxScore, Integer sortOrder) {}
